+++ packages/time-sync/src/TimeSync.ts
@@ -1,5 +1,4 @@
 import { ReadonlyDate } from "./ReadonlyDate";
-import type { Writeable } from "./utilities";
 
 /**
  * A collection of commonly-needed intervals (all defined in milliseconds).
@@ -30,7 +29,8 @@ export const refreshRates = Object.freeze({
 export interface Configuration {
 	/**
 	 * Indicates whether the TimeSync instance should be frozen for Snapshot
-	 * tests.
+	 * tests. Highly encouraged that you use this together with
+	 * `initialDate`.
 	 *
 	 * Defaults to false.
 	 */
@@ -44,7 +44,7 @@ export interface Configuration {
 	 * subscription, this minimum will be used instead.
 	 *
 	 * It is highly recommended that you only modify this value if you have a
-	 * good reason. Updating this value to be too low and make the event loop
+	 * good reason. Updating this value to be too low can make the event loop
 	 * get really hot and really tank performance elsewhere in an application.
 	 *
 	 * Defaults to 200ms.
@@ -55,9 +55,11 @@ export interface Configuration {
 	 * Indicates whether the same `onUpdate` callback (by reference) should be
 	 * called multiple time if registered by multiple systems.
 	 *
-	 * Defaults to true. If this value is flipped to false, each onUpdate
-	 * callback will receive the subscription context for the FIRST subscriber
-	 * that registered the onUpdate callback.
+	 * If this value is flipped to false, each onUpdate callback will receive
+	 * the subscription context for the FIRST subscriber that registered the
+	 * onUpdate callback.
+	 *
+	 * Defaults to true.
 	 */
 	readonly allowDuplicateOnUpdateCalls: boolean;
 }
@@ -66,16 +68,6 @@ export interface Configuration {
  * The set of options that can be used to instantiate a TimeSync.
  */
 export interface InitOptions extends Configuration {
-	/**
-	 * Indicates whether the TimeSync instance should be frozen for snapshot
-	 * tests. Highly encouraged that you use this together with
-	 * `initialDate`.
-	 *
-	 *  Defaults to false.
-	 */
-	// Duplicated property to override the LSP comment
-	readonly freezeUpdates: boolean;
-
 	/**
 	 * The Date object to use when initializing TimeSync to make the
 	 * constructor more pure and deterministic.
@@ -126,6 +118,14 @@ export interface Snapshot {
 	readonly date: ReadonlyDate;
 
 	/**
+	 * The monotonic milliseconds that elapsed between the TimeSync being
+	 * instantiated and the last update being dispatched.
+	 *
+	 * Will be null if no updates have ever been dispatched.
+	 */
+	readonly lastUpdatedAtMs: number | null;
+
+	/**
 	 * The number of subscribers registered with TimeSync.
 	 */
 	readonly subscriberCount: number;
@@ -139,18 +139,21 @@ export interface Snapshot {
 
 /**
  * An object with information about a specific subscription registered with
- * TimeSync.
- *
- * For performance reasons, this object has ZERO readonly guarantees enforced at
- * runtime. A few properties are flagged as readonly at the type level, but
- * misuse of this value has a risk of breaking a TimeSync instance's internal
- * state. Proceed with caution.
+ * TimeSync. The entire context is frozen at runtime.
  */
 export interface SubscriptionContext {
 	/**
-	 * The interval that the subscription was registered with.
+	 * A reference to the TimeSync instance that the subscription was registered
+	 * with.
 	 */
-	readonly targetRefreshIntervalMs: number;
+	readonly timeSync: TimeSync;
+
+	/**
+	 * The effective interval that the subscription is updating at. This may be a
+	 * value larger than than the target refresh interval, depending on whether
+	 * TimeSync was configured with a minimum refresh value.
+	 */
+	readonly refreshIntervalMs: number;
 
 	/**
 	 * The unsubscribe callback associated with a subscription. This is the same
@@ -159,31 +162,10 @@ export interface SubscriptionContext {
 	readonly unsubscribe: () => void;
 
 	/**
-	 * A timestamp of when the subscription was first set up.
+	 * The monotonic milliseconds that elapsed between the TimeSync being
+	 * instantiated and the subscription being registered.
 	 */
-	readonly registeredAt: ReadonlyDate;
-
-	/**
-	 * A reference to the TimeSync instance that the subscription was registered
-	 * with.
-	 */
-	readonly timeSync: TimeSync;
-
-	/**
-	 * Indicates whether the subscription is still live. Will be mutated to be
-	 * false when a subscription is
-	 */
-	isSubscribed: boolean;
-
-	/**
-	 * Indicates when the last time the subscription had its explicit interval
-	 * "satisfied".
-	 *
-	 * For example, if a subscription is registered for every five minutes, but
-	 * the active interval is set to fire every second, you may need to know
-	 * which update actually happened five minutes later.
-	 */
-	intervalLastFulfilledAt: ReadonlyDate | null;
+	readonly registeredAtMs: number;
 }
 
 /**
@@ -203,8 +185,16 @@ interface TimeSyncApi {
 	 * intervals. Depending on how TimeSync is instantiated, it may choose to
 	 * de-duplicate these function calls on each round of updates.
 	 *
-	 * @throws {RangeError} If the provided interval is not either a positive
-	 * integer or positive infinity.
+	 * If a value of Number.POSITIVE_INFINITY is used, the subscription will be
+	 * considered "idle". Idle subscriptions cannot trigger updates on their
+	 * own, but can stay in the loop as otherupdates get dispatched from via
+	 * other subscriptions.
+	 *
+	 * Consider using the refreshRates object from this package for a set of
+	 * commonly-used intervals.
+	 *
+	 * @throws {RangeError} If the provided interval is neither a positive
+	 * integer nor positive infinity.
 	 * @returns An unsubscribe callback. Calling the callback more than once
 	 * results in a no-op.
 	 */
@@ -232,9 +222,47 @@ interface TimeSyncApi {
 	clearAll: () => void;
 }
 
-/* biome-ignore lint:suspicious/noEmptyBlockStatements -- Rare case where we do
-   actually want a completely empty function body. */
-function noOp(..._: readonly unknown[]): void {}
+/**
+ * Even though both the browser and the server are able to give monotonic times
+ * that are at least as precise as a nanosecond, we're using milliseconds for
+ * consistency with useInterval, which cannot be more precise than a
+ * millisecond.
+ */
+function getMonotonicTimeMs(): number {
+	// If we're on the server, we can use process.hrtime, which is defined for
+	// Node, Deno, and Bun
+	if (typeof window === "undefined") {
+		const timeInNanoseconds = process.hrtime.bigint();
+		return Number(timeInNanoseconds / 1000n);
+	}
+
+	// Otherwise, we need to get the high-resolution timestamp from the browser.
+	// This value is fractional and goes to nine decimal places
+	const highResTimestamp = window.performance.now();
+	return Math.floor(highResTimestamp);
+}
+
+/**
+ * This function is just a convenience for us to sidestep some problems around
+ * TypeScript's LSP and Object.freeze. Because Object.freeze can accept any
+ * arbitrary type, it basically acts as a "type boundary" between the left and
+ * right sides of any snapshot assignments.
+ *
+ * That means that if you rename a property a a value that is passed to
+ * Object.freeze, the LSP can't auto-rename it, and you potentially get missing
+ * properties. This is a bit hokey, but because the function is defined strictly
+ * in terms of concrete snapshots, any value passed to this function won't have
+ * to worry about mismatches.
+ */
+function freezeSnapshot(snap: Snapshot): Snapshot {
+	if (!Object.isFrozen(snap.config)) {
+		Object.freeze(snap.config);
+	}
+	if (!Object.isFrozen(snap)) {
+		Object.freeze(snap);
+	}
+	return snap;
+}
 
 const defaultMinimumRefreshIntervalMs = 200;
 
@@ -274,6 +302,12 @@ const defaultMinimumRefreshIntervalMs = 200;
  * some parts of the screen.)
  */
 export class TimeSync implements TimeSyncApi {
+	/**
+	 * The monotonic time in milliseconds from when the TimeSync instance was
+	 * first instantiated.
+	 */
+	readonly #initializedAtMs: number;
+
 	/**
 	 * Stores all refresh intervals actively associated with an onUpdate
 	 * callback (along with their associated unsubscribe callbacks).
@@ -322,7 +356,8 @@ export class TimeSync implements TimeSyncApi {
 
 	/**
 	 * Used for both its intended purpose (creating interval), but also as a
-	 * janky version of setTimeout.
+	 * janky version of setTimeout. Also, all versions of setInterval are
+	 * monotonic, so we don't have to do anything special for it.
 	 *
 	 * There are a few times when we need timeout-like logic, but if we use
 	 * setInterval for everything, we have fewer IDs to juggle, and less risk of
@@ -354,63 +389,43 @@ export class TimeSync implements TimeSyncApi {
 		this.#subscriptions = new Map();
 		this.#fastestRefreshInterval = Number.POSITIVE_INFINITY;
 		this.#intervalId = undefined;
+		this.#initializedAtMs = getMonotonicTimeMs();
 
-		// Not defined inline to avoid wonkiness that Object.freeze introduces
-		// when you rename a property on a frozen object
-		const initialSnapshot: Snapshot = {
+		let date: ReadonlyDate;
+		if (initialDate instanceof ReadonlyDate) {
+			date = initialDate;
+		} else if (initialDate instanceof Date) {
+			date = new ReadonlyDate(initialDate);
+		} else {
+			date = new ReadonlyDate();
+		}
+
+		this.#latestSnapshot = freezeSnapshot({
+			date,
 			subscriberCount: 0,
-			date: initialDate ? new ReadonlyDate(initialDate) : new ReadonlyDate(),
-			config: Object.freeze({
+			lastUpdatedAtMs: null,
+			config: {
 				freezeUpdates,
 				minimumRefreshIntervalMs,
 				allowDuplicateOnUpdateCalls,
-			}),
-		};
-		this.#latestSnapshot = Object.freeze(initialSnapshot);
-	}
-
-	#setSnapshot(update: Partial<Snapshot>): boolean {
-		const { date, subscriberCount, config } = this.#latestSnapshot;
-		if (config.freezeUpdates) {
-			return false;
-		}
-
-		// Avoiding both direct property assignment or spread syntax because
-		// Object.freeze causes weird TypeScript LSP issues around assignability
-		// where trying to rename a property. If you rename a property on a
-		// type, it WON'T rename the runtime properties. Object.freeze
-		// introduces an extra type boundary that break the linking
-		const updated: Snapshot = {
-			// Always reject any new configs because trying to remove them at
-			// the type level isn't worth it for an internal implementation
-			// detail
-			config,
-			date: update.date ?? date,
-			subscriberCount: update.subscriberCount ?? subscriberCount,
-		};
-
-		this.#latestSnapshot = Object.freeze(updated);
-		return true;
+			},
+		});
 	}
 
 	#notifyAllSubscriptions(): void {
 		// It's more important that we copy the date object into a separate
 		// variable here than normal, because need make sure the `this` context
 		// can't magically change between updates and cause subscribers to
-		// receive different values (e.g., one of the subscribers calls the
-		// invalidate method)
+		// receive different values
 		const { date, config } = this.#latestSnapshot;
 
-		// We still need to let the logic go through if the current fastest
-		// interval is Infinity, so that we can support letting any arbitrary
-		// consumer invalidate the date immediately
 		const subscriptionsPaused =
-			config.freezeUpdates || this.#subscriptions.size === 0;
+			config.freezeUpdates ||
+			this.#subscriptions.size === 0 ||
+			this.#fastestRefreshInterval === Number.POSITIVE_INFINITY;
 		if (subscriptionsPaused) {
 			return;
 		}
-
-		const dateTime = date.getTime();
 
 		/**
 		 * Two things:
@@ -431,34 +446,20 @@ export class TimeSync implements TimeSyncApi {
 		 * to check on each iteration to see if we should continue.
 		 */
 		const subsBeforeUpdate = this.#subscriptions;
-		const entries = Array.from(subsBeforeUpdate);
-		outer: for (const [onUpdate, subs] of entries) {
-			// Even if duplicate onUpdate calls are disabled, we still need to
-			// iterate through everything and update any internal data. If the
-			// first context in a sub array gets removed by unsubscribing, we
-			// want what was the the second element to still be up to date
-			let shouldCallOnUpdate = true;
-			for (const context of subs) {
+		const localEntries = Array.from(subsBeforeUpdate);
+		outer: for (const [onUpdate, subs] of localEntries) {
+			for (const ctx of subs) {
 				// We're not doing anything more sophisticated here because
 				// we're assuming that any systems that can clear out the
 				// subscriptions will handle cleaning up each context, too
-				const wasCleared = subsBeforeUpdate.size === 0;
-				if (wasCleared) {
+				const wasClearedBetweenUpdates = subsBeforeUpdate.size === 0;
+				if (wasClearedBetweenUpdates) {
 					break outer;
 				}
 
-				const comparisonDate =
-					context.intervalLastFulfilledAt ?? context.registeredAt;
-				const isIntervalMatch =
-					dateTime - comparisonDate.getTime() >=
-					context.targetRefreshIntervalMs;
-				if (isIntervalMatch) {
-					context.intervalLastFulfilledAt = date;
-				}
-
-				if (shouldCallOnUpdate) {
-					onUpdate(date, context);
-					shouldCallOnUpdate = config.allowDuplicateOnUpdateCalls;
+				onUpdate(date, ctx);
+				if (!config.allowDuplicateOnUpdateCalls) {
+					continue outer;
 				}
 			}
 		}
@@ -473,56 +474,65 @@ export class TimeSync implements TimeSyncApi {
 	 * is one of them.
 	 */
 	readonly #onTick = (): void => {
-		// Defensive step to make sure that an invalid tick wasn't started
 		const { config } = this.#latestSnapshot;
 		if (config.freezeUpdates) {
+			// Defensive step to make sure that an invalid tick wasn't started
 			clearInterval(this.#intervalId);
 			this.#intervalId = undefined;
 			return;
 		}
 
-		const wasChanged = this.#setSnapshot({ date: new ReadonlyDate() });
-		if (wasChanged) {
-			this.#notifyAllSubscriptions();
-		}
+		// onTick is expected to be called in response to monotonic time changes
+		// (either from calculating them manually to decide when to call onTick
+		// synchronously or from letting setInterval handle the calls). So while
+		// this edge case should basically be impossible, we need to make sure that
+		// we always dispatch a date, even if its time is exactly the same.
+		this.#latestSnapshot = freezeSnapshot({
+			...this.#latestSnapshot,
+			date: new ReadonlyDate(),
+			lastUpdatedAtMs: getMonotonicTimeMs() - this.#initializedAtMs,
+		});
+		this.#notifyAllSubscriptions();
 	};
 
 	#onFastestIntervalChange(): void {
 		const fastest = this.#fastestRefreshInterval;
-		const { date, config } = this.#latestSnapshot;
+		const { lastUpdatedAtMs, config } = this.#latestSnapshot;
+
 		const updatesShouldStop =
-			config.freezeUpdates || fastest === Number.POSITIVE_INFINITY;
+			config.freezeUpdates ||
+			this.#subscriptions.size === 0 ||
+			fastest === Number.POSITIVE_INFINITY;
 		if (updatesShouldStop) {
 			clearInterval(this.#intervalId);
 			this.#intervalId = undefined;
 			return;
 		}
 
-		const elapsed = new ReadonlyDate().getTime() - date.getTime();
+		const newTime = getMonotonicTimeMs();
+		const elapsed = newTime - (lastUpdatedAtMs ?? this.#initializedAtMs);
 		const timeBeforeNextUpdate = fastest - elapsed;
 
-		// Clear previous interval sight unseen just to be on the safe side
+		// Clear previous interval no matter what just to be on the safe side
 		clearInterval(this.#intervalId);
 
 		if (timeBeforeNextUpdate <= 0) {
-			const wasChanged = this.#setSnapshot({ date: new ReadonlyDate() });
-			if (wasChanged) {
-				this.#notifyAllSubscriptions();
-			}
+			this.#onTick();
 			this.#intervalId = setInterval(this.#onTick, fastest);
 			return;
 		}
 
 		// Most common case for this branch is the very first subscription
 		// getting added, but there's still the small chance that the fastest
-		// interval could change right after an update got flushed
+		// interval could change right after an update got flushed, so there would
+		// be zero elapsed time to worry about
 		if (timeBeforeNextUpdate === fastest) {
 			this.#intervalId = setInterval(this.#onTick, timeBeforeNextUpdate);
 			return;
 		}
 
-		// Otherwise, use interval as pseudo-timeout, and then go back to using
-		// it as a normal interval afterwards
+		// Otherwise, use setInterval as pseudo-timeout to resolve the remaining
+		// time as a one-time update, and then go back to using normal intervals
 		this.#intervalId = setInterval(() => {
 			clearInterval(this.#intervalId);
 
@@ -543,14 +553,13 @@ export class TimeSync implements TimeSyncApi {
 			return;
 		}
 
-		const prevFastest = this.#fastestRefreshInterval;
-		let newFastest = Number.POSITIVE_INFINITY;
-
 		// This setup requires that every interval array stay sorted. It
 		// immediately falls apart if this isn't guaranteed.
-		for (const entries of this.#subscriptions.values()) {
+		const prevFastest = this.#fastestRefreshInterval;
+		let newFastest = Number.POSITIVE_INFINITY;
+		for (const contexts of this.#subscriptions.values()) {
 			const subFastest =
-				entries[0]?.targetRefreshIntervalMs ?? Number.POSITIVE_INFINITY;
+				contexts[0]?.refreshIntervalMs ?? Number.POSITIVE_INFINITY;
 			if (subFastest < newFastest) {
 				newFastest = subFastest;
 			}
@@ -563,14 +572,10 @@ export class TimeSync implements TimeSyncApi {
 	}
 
 	subscribe(options: SubscriptionInitOptions): () => void {
-		const { config } = this.#latestSnapshot;
-		if (config.freezeUpdates) {
-			return noOp;
-		}
-
 		// Destructuring properties so that they can't be fiddled with after
 		// this function call ends
 		const { targetRefreshIntervalMs, onUpdate } = options;
+		const { minimumRefreshIntervalMs } = this.#latestSnapshot.config;
 
 		const isTargetValid =
 			targetRefreshIntervalMs === Number.POSITIVE_INFINITY ||
@@ -582,80 +587,80 @@ export class TimeSync implements TimeSyncApi {
 			);
 		}
 
-		// Have to define this as a writeable to avoid a chicken-and-the-egg
-		// problem for the unsubscribe callback
-		const context: Writeable<SubscriptionContext> = {
-			isSubscribed: true,
+		const subsOnSetup = this.#subscriptions;
+		let subscribed = true;
+		const ctx: SubscriptionContext = {
 			timeSync: this,
-			unsubscribe: noOp,
-			registeredAt: new ReadonlyDate(),
-			intervalLastFulfilledAt: null,
-			targetRefreshIntervalMs: Math.max(
-				config.minimumRefreshIntervalMs,
+			registeredAtMs: getMonotonicTimeMs() - this.#initializedAtMs,
+			refreshIntervalMs: Math.max(
+				minimumRefreshIntervalMs,
 				targetRefreshIntervalMs,
 			),
+
+			unsubscribe: () => {
+				// Not super conventional, but basically using try/finally as a form of
+				// Go's defer. There are a lot of branches we need to worry about for
+				// the unsubscribe callback, and we need to make sure we flip subscribed
+				// to false after each one
+				try {
+					if (!subscribed || this.#subscriptions !== subsOnSetup) {
+						return;
+					}
+					const contexts = subsOnSetup.get(onUpdate);
+					if (contexts === undefined) {
+						return;
+					}
+					const filtered = contexts.filter(
+						(c) => c.unsubscribe !== ctx.unsubscribe,
+					);
+					if (filtered.length === contexts.length) {
+						return;
+					}
+
+					const dropped = Math.max(0, this.#latestSnapshot.subscriberCount - 1);
+					this.#latestSnapshot = freezeSnapshot({
+						...this.#latestSnapshot,
+						subscriberCount: dropped,
+					});
+
+					if (filtered.length > 0) {
+						// No need to sort on removal because everything gets sorted as
+						// it enters the subscriptions map
+						subsOnSetup.set(onUpdate, filtered);
+					} else {
+						subsOnSetup.delete(onUpdate);
+					}
+
+					this.#updateFastestInterval();
+				} finally {
+					subscribed = false;
+				}
+			},
 		};
+		Object.freeze(ctx);
 
-		// Not reading from context value to decide whether to bail out of
-		// unsubscribes in off chance that outside consumer accidentally mutates
-		// the value
-		let subscribed = true;
-		const subsOnSetup = this.#subscriptions;
-		const unsubscribe = (): void => {
-			if (!subscribed || this.#subscriptions !== subsOnSetup) {
-				context.isSubscribed = false;
-				subscribed = false;
-				return;
-			}
-
-			const contexts = subsOnSetup.get(onUpdate);
-			if (contexts === undefined) {
-				return;
-			}
-			const filtered = contexts.filter((e) => e.unsubscribe !== unsubscribe);
-			if (filtered.length === contexts.length) {
-				return;
-			}
-
-			if (filtered.length === 0) {
-				subsOnSetup.delete(onUpdate);
-				this.#updateFastestInterval();
-			} else {
-				// No need to sort on removal because everything gets sorted as
-				// it enters the subscriptions map
-				subsOnSetup.set(onUpdate, filtered);
-			}
-
-			void this.#setSnapshot({
-				subscriberCount: Math.max(0, this.#latestSnapshot.subscriberCount - 1),
-			});
-
-			context.isSubscribed = false;
-			subscribed = false;
-		};
-		context.unsubscribe = unsubscribe;
-
-		let contexts: SubscriptionContext[];
-		if (this.#subscriptions.has(onUpdate)) {
-			const prev = this.#subscriptions.get(onUpdate) as SubscriptionContext[];
-			contexts = [...prev];
+		// We need to make sure that each array for tracking subscriptions is
+		// readonly, and because dispatching updates should be far more common than
+		// adding subscriptions, we're placing the immutable copying here to
+		// minimize overall pressure on the system.
+		let newContexts: SubscriptionContext[];
+		const prevContexts = subsOnSetup.get(onUpdate);
+		if (prevContexts !== undefined) {
+			newContexts = [...prevContexts, ctx];
 		} else {
-			contexts = [];
-			subsOnSetup.set(onUpdate, contexts);
+			newContexts = [ctx];
 		}
 
-		subsOnSetup.set(onUpdate, contexts);
-		contexts.push(context);
-		contexts.sort(
-			(e1, e2) => e1.targetRefreshIntervalMs - e2.targetRefreshIntervalMs,
-		);
+		subsOnSetup.set(onUpdate, newContexts);
+		newContexts.sort((c1, c2) => c1.refreshIntervalMs - c2.refreshIntervalMs);
 
-		void this.#setSnapshot({
+		this.#latestSnapshot = freezeSnapshot({
+			...this.#latestSnapshot,
 			subscriberCount: this.#latestSnapshot.subscriberCount + 1,
 		});
 
 		this.#updateFastestInterval();
-		return unsubscribe;
+		return ctx.unsubscribe;
 	}
 
 	getStateSnapshot(): Snapshot {
@@ -670,17 +675,15 @@ export class TimeSync implements TimeSyncApi {
 		// As long as we clean things the internal state, it's safe not to
 		// bother calling each unsubscribe callback. Not calling them one by
 		// one actually has much better time complexity
-		for (const subArray of this.#subscriptions.values()) {
-			for (const ctx of subArray) {
-				ctx.isSubscribed = false;
-			}
-		}
-
 		this.#subscriptions.clear();
 
 		// We swap the map out so that the unsubscribe callbacks can detect
 		// whether their functionality is still relevant
 		this.#subscriptions = new Map();
-		void this.#setSnapshot({ subscriberCount: 0 });
+
+		this.#latestSnapshot = freezeSnapshot({
+			...this.#latestSnapshot,
+			subscriberCount: 0,
+		});
 	}
 }