+++ packages/time-sync/CHANGELOG.md
@@ -1,8 +1,40 @@
 # @buenos-nachos/time-sync
 
+## 0.5.0
+
+### Breaking Changes
+
+- c3986e9: revamped all state management and APIs to be based on monotonic time
+- c3986e9: Removed `registeredAt` and `intervalLastFulfilledAt` properties from `SubscriptionContext` and added monotonic `registeredAtMs`
+- c3986e9: Added monotonic `lastUpdatedAt` property to `Snapshot` type.
+
+## 0.4.1
+
+### Patch Changes
+
+- 5f37f1a: refactored class to remove private setSnapshost method
+
+## 0.4.0
+
+### Breaking Changes
+
+- 663479e: Removed `isSubscribed` property from context and made all other context properties readonly.
+
+## 0.3.2
+
+### Patch Changes
+
+- b8fbaf8: cleanup up comments and types for exported class, methods, and types.
+
+## 0.3.1
+
+### Patch Changes
+
+- 5fce018: switched internal implementations to use Date.now more often to reduce memory usage
+
 ## 0.3.0
 
-### Minor Changes
+### Breaking Changes
 
 - 122f6c1: Updated `SubscriptionContext.timeSync` type to be readonly and non-nullable, and renamed `SubscriptionContext.isLive` to `SubscriptionContext.isSubscribed`.